Fix stripe payment rounding bug
Stripe payment backend is rounding the total before multiplying by 100,
which results in an incorrect charge.

This commit fixes the bug, as well as adding a test to illustrate the
problem.

This commit also makes a number of other changes so that we can run
this type of test:

- Adds mock and stripe as dependencies for testing (edits to travis.yml)
- Adds STRIPE_API_KEY to local_settings.py.template so travis tests
  will work properly

Also documents how to run Cartridge unit tests in the readme.